stt/requirements.txt
vosk==0.3.44
sounddevice==0.5.2
pyglet
tts
cffi==1.17.1
tqdm==4.67.1
python-dotenv==1.1.0
httpx==0.24.1
# srt==3.5.3
# websockets==15.0.1
